requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==1.0.0
Django==3.2.15
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.1
idna==3.3
mssql-django==1.1.3
psycopg2==2.9.3
pyodbc==4.0.34
pytz==2022.2.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.12
whitenoise==5.3.0
